Suppress warnings, change output plot name, auto-size plot
This commit changes the filename of the plot of the arrays to
have a "_flagCosmicRays" suffix. It also automatically adjusts
the figure size base on the aspect ratio of the images, which
allows for both full-frame and center-frame images to be
displayed without significant extraneaous whitespace. Finally,
it suppresses "divide by zero" and "invalid value encountered"
warnings from NumPy that show up when saving the figure if
either the object or sky line models are not present.