Fix additional issues, polishing for detemineSlitEdges()
This commit fixes several issues found with various datasests.
Edges expected to be off the side of the detector are no
longer matched to peaks in the image, to avoid spurious matches
with fake peaks (most notable in some GNIRS 'Long' camera data).

Fitting of edges now starts at the row with the highest median
flux, uses the median flux of each row as the weight for that row,
and can handle areas up to 160 pixels where the illumination drops
to the background level. The new min_peak_value of trace_lines()
is used with a value of 5 times the read noise to avoid picking
up peaks in the noise where the illumination drops to the back-
ground level. Overall this makes edge fitting much more robust.

In addition, determineSlitEdges() now takes edges1 and edges2
parameters where the expected edge positions (in pixels on the
detector) can be given. This allows for situations where the
illuminated region breadth is more than 2% different from the
expected value based on the MDF/LUT.