Changes related to slit rectification
Abstract a create_distortion_model() function from code in
distortionCorrect(). Change determineSlitEdges() to call this new
function to created a "rectified" transform (added to the WCS)
which transforms the illuminated region edges to be straight.

Also tweak attachWavelengthSolution to handle have this new
"rectified" frame/transform in the WCS, instead of assuming the
WCS is the default.y