Use log-weighting for fit, add order of fit as parameter
This commit adds the `spectral_order` parameter to determineSlitEdges,
which controls the order of the fit used for the edges of the
illuminated region in the flat. Additionally, the weights used in
the fitting process were changed from the square root of the summed
flux in each row/column (in the spatial direction) to the log,
which drastically reduces the range of weight values and ensures
weaker rows which might otherwise not be fitted are still included
in the fit.

With this change, one of the test cases for test_determine_slit_edges
no longer passed, due to now fitting points which were being
rejected using the square-root weighting system. The coefficients
for this test to compare to were regenerated, along with having it
use a 4th-order fit since that was necessary for a better RMS in
this case.