Minor comment improvements
Fix all linear cross-order terms in the spatial direction in fitting

Before this commit only the lowest degree cross-order term was fixed,
which resulted in the edge of the distortion-corrected region being
rectified correctly, but not necessarily the rest of the region:
object traces might be over- or under-rectified, for instance.

This change fixes all higher-degree cross-order terms as well, which
results in the distortion being a fixed linear amount in the spatial
direction.