Bunch of work to fit sky lines and actually shift wavelength scale
With this release the wavelength scale is actually shifted by the
amount found in fitting the sky lines. A few parameters were
added to adjustWavelengthZeroPoint for this purpose.

read_lines() in gempy/library/wavecal.py now also checks if the
weights it gets from np.genfromtxt() are all NaNs, and sets the
weights to None if so (same as if nothing was returned).