Add 'skylines' and 'loglevel' parameters to get_all_input_data()
'skylines' just controls what to call the reference linelist lines
in output. 'loglevel' controls at what level to print to the log
(existing log.warning calls are left untouched, however). The
default is 'stdinfo', which acts as previously, however, if
suppressing output in the terminal (visual) log is desired it
can be called as 'fullinfo' or 'debug', so that it only appears
if the log is set to show those levels.

Add documentation for 'loglevel' parameter