Define bounds on c2 in wavecal fitting model within perform_piecewise_fit
This commit changes the behavior the calculation of bounds on the c2
parameter of the Chebyshev1D model used for wavelength calibration as
originally defined in commit 142aa74. For some unknown reason, passing a
value through to perform_piecewise_fit() caused a single case in the GMOS
regression test test_regression_determine_wavelength_solution to fail,
though *only* on the Jenkins server, with Python 3.10, and not when just
that single case was run. The ultimate cause of this bug was never found,
as it was discovered that defining the bound within perform_piecewise_fit()
allowed the test to pass. This commit changes the behavior to do that,
simultaneously removing the defining of 'bound' in primitives_spect.py.