Remove step of matching peaks to skylines before fitting shift
Matching peaks to skylines uses the stochastic differential
evolution minimization method, which was seen to produce
essentially random results on spectra with many closely-spaced
sky lines. It turned out to be unnecessary for fitting the shift,
which worked repeatably and reliably without it. (Removing it also
greatly sped up adjustWavelengthZeroPoint() as it added a lot
of time to the overall runtime.)

Also adds a small 'verbose' option the fitting procedure.