More informative error message when unable to find sky lines in spectrum
Added an error check when calling get_all_input_data() and it returns
a ValueError due to being unable to find sky lines. This can happen
for (at least) two reasons: in the first case, the object aperture
may be located in the central columns/rows used for extracting the sky
spectrum; the error message points out this possibility and suggests
using the `center` parameter to change the column/row being extracted.

The second possibility is that the SNR for the observation is too low
to allow lines to be found. This is impossible to tell apart from the
first case at the level of the primitive, but should hopefully be less
common.