Overhaul of testing documentation
Adds a paragraph with instructions on installing pytest_dragons.

Adds a table listing fixtures in astrodata/test.py.

Removes an old example of a test function which wasn't up-to-date, replaces it
with three new examples that (hopefully) better illustrate the principles and
fixtures involved.

Some minor changes and additional examples of directory structures.