Initial commit of files for GNIRS XD support
Adds some primitives for GNIRS XD data, specifically relating to
mask-definition of the illuminated regions on the detector. A
new XD-specific addMDF() uses a look-up file added by this commit
to add a table of values defining the centers and widths of the
various orders.

The keys in maskdb.py have also been extended to support unique
combinations for XD and IFU observing modes in addition to LS.