Update determineSlitEdges() handle XD data
This initial commit add support for XD data to determineSlitEdges()
by allowing it to fit more than a single pair of slit edges in an
image. The SLITEDGE table returned contains the model coefficients
for each of the slits, numbered such that each pair can be kept
track of.

There are also some tweaks to the formatting of log output, to
make it easier to keep track of when fitting multiple slits in the
same image.

Update _get_slit_edge_estimates() to handle multiple slits