Move cutSlits() primitive to mode-specific files
This commit moves the code providing the main functionality of
cutSlits() to a private _cut_slits() method in Spect. It also
adds a cutSlits() primitive in primitives_crossdispersed.py
which handles modifications to the WCS resulting from cutting
slits, and calls the new _cut_slits() function.

The reason for this change is the aforementioned modification
of the WCS that needs to happen for each new extension. The
exact change are certainly instrument-specific (and likely mode-
specific even within instruments), so need to be handle on a
per-instrument basis. This was done by having instrument-specific
cutSlits() which super()-call the Spect version of cutSlits();
however, since that parent version would never be complete when
called on its own, it was decided to move its code (the actual
"slit cutting" code) into a private method which the various
instrument specific cutSlits() could call (only GNIRS so far).

Initial commit of file used by instrument-specific cutSlits()

Add key to order-specific WCS modification information