Refactor trace_lines() use in determineDistortion() to fix tests
A previous commit changed the call to tracing.trace_lines() in
deteremineDistortion() so that instead of passing all the peaks
found in initial_peaks at once, it passed them one at a time and
concatenated the resulting the traced coordinates together. This
is necessary for cross-dispersed and other spectra where the aperture
isn't straight and the illuminated region doesn't cover the entire
width of the spatial direction, requiring a separate "start"
location for each line.

This works, but is quite a bit slower as the number of lines
increases, and also contained a small bug with hardcoding the
value of min_line_length. This commit fixes that bug (so that
the tests pass again), and also introduces an if-clause that
sends the peaks all together for longslit (or other "straight"
slits), but one at a time for cross-dispersed (or other future
"curved" slits). Both methods (testably) produce the same
results for longslit data, but the former is much faster, so
we might as well use it where we can.