Fix tests for Astrodata tags and descriptors
In September 2023 it was discovered that the tests for Astrodata
tags and descriptors were not being automatically discovered by
Pytest (due to their filenames not starting with 'test_'). From
the fact that they were still using hardcoded test data filepaths
and hadn't been updated in three years, it is probable that they
hadn't been running for at least a few years. When run manually,
they contained numerous failures.

This commit fixes issues found in the tests. In particular, it:
1) Changes to downloading raw data (the vast majority of files)
   from the archive, and storing ~10 processed files in the
   standard $DRAGONS_TEST directory.
2) Updates reference values of tags and descriptors where they
   were out of date.
3) Fixes a few bugs (in TRECS, NIRI, and NIFS) which had gone
   unnoticed due to tests not running.

Update group_id for GMOS test file