Javascript fix to link css to bokeh models
This adds a new function that is called when the window is loaded in
the browser. It recursively finds all shadow-roots in the generate
HTML and makes sure they know about the style sheets.

There is a small impact on page loading time (about 0.5 sec), and as
it is now it will render the page briefly without css and then apply it.
I'm exploring ways to mitigate that.

This does not fix all of the layout issues, but the remaining ones are not
handled by the stylesheet and should be manageable with the
migration guide.