Adds working inline stylesheet via bokeh
This fix wasn't working before because I had missed a single
Div or something while modifying it the first time around.

GlobalImportedStylesheet does not work because it doesn't get
referenced by bokeh properly (it claims to be "referenced once"
but the output HTML does not have the styles applied to any
of the shadow roots even though they're aware of it).

So, ImportedStyleSheet seems to be the way to go. The same sizing
issues that occur when manually applying stylesheet links through
javascript (which this is *not* equivalent to), which are likely because
of changes to sizing_mode/(height|width)_policy.

The fixes here are only applied to the interactive and fit1d libraries.