Add primitives for using pinholes for slit rectification models
This commit adds two new primitives, attachPinholeModel and
tracePinholeApertures, which are used to create a processed
pinhole file, generate a slit rectification model from it, and
attach that model to arcs and science frames for GNIRS XD data.

It also adds a function _attach_slit_rectification_model, which
copies a slit rectification from one file (intended to be a
processed pinhole) to another (an arc or science frame).

Includes some minor fixes in _cut_slits as well.