Add warnings for unused method kwargs
These are always enabled in this method, and the
kwargs are only here to match the signature of
Fit1DPanel.build_figures. Since these are default True,
it only warns if a user attempts to disable them.

The better solution is probably to use them in the long run, but
right now I'm just preserving current behavior.