Generate reference using sphinx-autoapi
sphinx-autoapi automatically generates a complete reference
by *statically* parsing the source files in our code. This is a new,
straightforward way to interact with existing and new documentation
otherwise only accessible through docstrings.

This does not propogate to any other pages in the documentation not
generated at the top-level (e.g., the astrodata/recipe manuals). But
these are still included in the reference API regardless.

This also allows smart linking to the API reference from within
docstrings. e.g., if one wants to reference the AstroData class,
you can just do the following:
    :py:class:`~astrodata.core.AstroData`

This creates a link to the reference entry for AstroData's class. If
desired it could be included using a replacement directive
    .. |AstroData| replace:: :py:class:`~astrodata.core.AstroData`

There are other options/configurations to streamline this a bit. This
commit also build fine, but there are some circular references and
duplicate object references that need to be either resolved or better
handled.

Obviously, there is no good way to implement this other than finding
references in the docstrings and changing them "manually" if they
don't properly link to the right place. I'm writing a script for the
astrodata package to maintain a list of these kinds of broken links
that could probably be applied to this. But, no ETC on that.