Fix negative peaks causing erroneously wide apertures
This commit should fix the issue of apertures sometimes being much
wider than they should be when found automatically. This happened
because the aperture width is determined as a fraction between a
putative aperture's peak and the adjacent minima. In some observations,
the removal of insignificant minima could lead to those "nearest
minima" being at the bottom of the negative troughs in IR imaging,
which are much further away than expected and which would pull the
aperture edges towards them.

To fix this, maxima below the median value of the profile are
merged, then any minima remaining below the median are set to
the median, thus removing negative values.