Add SPECORDR keyword to extensions when being cut out
For modes such as cross-dispersed where different orders of the same
slit appear on the detector, this commit adds a SPECORDR keyword
when those slits are cut out into their own extensions. This happens
in _cut_slits() in primitives_spec.py, and the keyword is copied to
the extensions created by gemini_tools.cut_to_match_auxiliary_data().