Update do_plots() to work with GNIRS data
Several fixes were made, including to handle vertically-dispersed
data and to replace a matplotlib colormap which had been deprecated
since version 2.3.1. Also handle the possibility of slashes in
the filename (composed of various information like filter, grating,
etc.) by replacing them with hyphens, as this was being interpreted
as a new directory level when saving.