Fill out predict_location() method of Trace
The idea behind this method is to predict, based on a user-specifiable
number of the last N points in a trace, where the next peak "should"
be found. This is mainly useful for traces that are non-vertical and
have been temporarily lost, with the need arising from fitting GNIRS
XD pinhole traces in the higher orders where the traces are obviously
curved, and tightly packed. (The pinholes must be traced to provide
a rectification model in the first place, which is why they can't be
rectified first and then traced.)

The method uses a linear model to extrapolate, so will likely only
be useful for relatively short extrapolation distances when tracing
non-straight slits.

Minor docstring changes, add missing braces in f-string.