Overhaul of log output for determineSlitEdges()
Changes most log output from determineSlitEdges() to be clearer
and show more information such as which edge is being referred to.
Also uses 1-indexing when referencing rows/columns and slits in
user-facing information, and uses "orders" instead of slits for
cross-dispersed data.

Also updated a few comments that were out of date.