Refactor trace_lines() to return list of Trace objects
Moves the process of creating in_coords and ref_coords (which
were previously created in trace_lines()) out to the calling
code, and have trace_lines() simply return a list of Trace
objects. (Or an empty list if none of the peaks it was given
could be recovered, which is mainly intended for when fitting
lists of arc/sky lines one by one along a non-straight slit.)

Also introduces the reverse_returned_coords parameter upon
initializing a Trace object. Since tracing happens vertically,
Traces.points is a series of (y, x) points. Since it may be
desirable to have points in (x, y) format, the new parameter
causes points returned by the methos inputs_coordinates() and
reference_coordinates() to be reversed.