Abstract tracing single aperture into trace_aperture() function
Removes the separate calls to trace_lines() from the interactive
and non-interactive branches of traceApertures() and abstracts
them into a single function tracing.trace_aperture(). The original
driver for this was because the two branches were calculating
their starting locations differently, leading to different results
(which is a litte worrisome, but another matter).