Put flatCorrect() after stackFrames() to avoid a crash
flatCorrect() attaches a SLITEDGE table which is needed in
determineWavelengthSolution() (to get a function representing the
center of the slit for average_along_slit()), but stackFrames()
doesn't transmit it. There shouldn't be any real impact on reduction
doing it either way, so until stackFrames() has a consistent way of
handling tables just move flatCorrect() after it.