Dynamically set central wavelength in each XD order
This commit changes cutSlits() so that instead of simply adding
a pre-defined central wavelength to the WAVE model of the WCS
from order_XD.py, it instead adds a central wavelength calculated
from the central wavelength chosen for the observation. This
central wavelength is chosen in one specific order, and with that
knowledge the central wavelength of all the other orders can be
calculated using the following formula:

  order_X * central_wavelength_X = order_Y * central_wavelength_Y

This formula should hold for any cross-dispersed instrument.