Fix find_outliers() when given repeated small number of values, add test
find_outliers() was written with the intent that it would be dealing
with noisy values, so a set of points in the input data was taken.
In some cases where the input has multiple repeated zeros, this
can produce incorrect results, for instance, in the sequence
[0, 17, 0, 0] all of the values end up masked, instead of just
the 17. This commit changes the code to use a list to preserve
all input values. It also adds a test with this edge case and
a few other distributions tested.