Update how dispersion is added in cutSlits()
This commit removes the geminidr/gnirs/lookups/orders_XD.py file
and changes how the initial dispersion is added to the WAVE model
in the WCS of each order cut out in cutSlits(). The dispersion
changes by order the same way the central wavelength does, so we
only need an initial dispersion and the order it applies to to
calculate it for all the other orders. Dispersion information for
order 3-6 is found in gemini_instruments.gnirs.lookup.py, so this
commit gets it for order 3 and then calculates it for each other
spectral order analogously to the central wavelength.